Added Question Listing in Question Module

+++ src/app/core/layout/sidenav/sidenav.component.ts
@@ -4,14 +4,13 @@ import { LoginService } from 'src/app/shared/services/login.service';
 @Component({
   selector: 'app-sidenav',
   templateUrl: './sidenav.component.html',
-  styleUrls: ['./sidenav.component.css']
+  styleUrls: ['./sidenav.component.css'],
 })
 export class SidenavComponent implements OnInit {
-
   navData = NAVBAR_DATA;
   collapsed = true;
   screenWidth = 0;
-  loginUserName:string|undefined;
+  loginUserName: string | undefined;
 
   toggleCollapse(): void {
     this.collapsed = !this.collapsed;
@@ -20,7 +19,7 @@ export class SidenavComponent implements OnInit {
     this.collapsed = false;
   }
 
-  constructor(private loginService: LoginService){}
+  constructor(private loginService: LoginService) {}
 
   ngOnInit(): void {
     let loginData = this.loginService.getUserDetailsFromLocalStorage();
@@ -28,27 +27,29 @@ export class SidenavComponent implements OnInit {
   }
 
   onLogout() {
-    if(confirm("Are u sure ?"))
-      this.loginService.removeTokenFromStorage();
+    if (confirm('Are u sure ?')) this.loginService.removeTokenFromStorage();
   }
-
 }
 
 const NAVBAR_DATA = [
   {
     routerLink: '/exam-portal/dashboard',
-    img:'../../../../assets/images/home.svg',
-    label: 'Dashhboard'
+    img: '../../../../assets/images/home.svg',
+    label: 'Dashhboard',
   },
   {
     routerLink: '/exam-portal/admin/exam',
-    img:'../../../../assets/images/book.svg',
-    label: 'Exam'
+    img: '../../../../assets/images/book.svg',
+    label: 'Exam',
   },
   {
     routerLink: '/exam-portal/admin/category',
-    img:'../../../../assets/images/category.svg',
+    img: '../../../../assets/images/category.svg',
     label: 'Category',
   },
-
-]
+  {
+    routerLink: '/exam-portal/admin/questions',
+    img: '../../../../assets/images/question.svg',
+    label: 'Questions',
+  },
+];
